requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
python-dotenv
google-generativeai
